requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
requests>=2.26.0
python-dateutil>=2.8.0
pytz>=2021.1
python-binance>=1.0.19
python-dotenv>=0.19.0
